feat: 스케줄 별 기수 전체 출석 현황 조회

+++ src/main/java/com/ddd/manage_attendance/domain/attendance/api/dto/AttendanceSummaryResponse.java
@@ -4,9 +4,9 @@ import com.ddd.manage_attendance.domain.attendance.domain.AttendanceSummary;
 import io.swagger.v3.oas.annotations.media.Schema;
 
 public record AttendanceSummaryResponse(
-        @Schema(description = "총 출석 일", example = "1") int totalAttended,
-        @Schema(description = "총 지각 일", example = "2") int totalLate,
-        @Schema(description = "총 결석 일", example = "3") int totalAbsent) {
+        @Schema(description = "총 출석 일", example = "1") long totalAttended,
+        @Schema(description = "총 지각 일", example = "2") long totalLate,
+        @Schema(description = "총 결석 일", example = "3") long totalAbsent) {
     public static AttendanceSummaryResponse from(final AttendanceSummary attendanceSummary) {
         return new AttendanceSummaryResponse(
                 attendanceSummary.getAttendedCount(),
