+++ src/main/java/com/ddd/manage_attendance/domain/attendance/domain/AttendanceSummary.java
@@ -1,3 +1,3 @@
 package com.ddd.manage_attendance.domain.attendance.domain;
 
-public record AttendanceSummary(int getAttendedCount, int getAbsentCount, int getLateCount) {}
+public record AttendanceSummary(long getAttendedCount, long getAbsentCount, long getLateCount) {}
